updated database to v3 and synchronised pictures api

+++ Backend/PicturesAPI/app/models/hair_style_link.py
@@ -4,7 +4,7 @@ Author: Diego C. <20026893@tafe.wa.edu.au>
 Created at: 28/09/2020 10:16 pm
 File: hair_style_link.py
 """
-from sqlalchemy import func, Column, String, Integer, DateTime, ForeignKey
+from sqlalchemy import func, Column, String, BIGINT, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 from app.database.db import Base
 
@@ -13,8 +13,8 @@ from .hair_style import HairStyle
 
 class HairStyleLinks(Base):
     __tablename__ = "hair_style_links"
-    id = Column(Integer, primary_key=True, index=True)
-    hair_style_id = Column(Integer, ForeignKey("hair_styles.id"), nullable=False)
+    id = Column(BIGINT, primary_key=True, index=True)
+    hair_style_id = Column(BIGINT, ForeignKey("hair_styles.id"), nullable=False)
     link_name = Column(String(128), nullable=False, server_default="** ERROR: missing category **")
     link_url = Column(String(512), nullable=False, server_default="** ERROR: missing category **")
     date_created = Column(DateTime(timezone=True), server_default=func.now())
